feat: ai api 결과 반환값 받기(#93)
