+++ src/main/java/capstone/safeat/filter/external/dto/FoodInferenceResponse.java
@@ -1,16 +1,13 @@
-package capstone.safeat.filter.external;
+package capstone.safeat.filter.external.dto;
 
 import capstone.safeat.category.domain.Category;
 import capstone.safeat.filter.domain.EstimatedFood;
+import capstone.safeat.filter.external.dto.FoodEstimateResponse.Ingredient;
 import java.util.List;
 
-public record FoodEstimateResponse(
-    String koreanName, List<Ingredient> ingredients, boolean isAmbiguous, boolean isFood
+public record FoodInferenceResponse(
+    String koreanName, List<Ingredient> ingredients
 ) {
-
-  public record Ingredient(String englishName) {
-
-  }
 
   public EstimatedFood toEstimateCategory() {
     final List<Category> categories = ingredients().stream()
@@ -21,9 +18,7 @@ public record FoodEstimateResponse(
 
     return EstimatedFood.builder()
         .categories(categories)
-        .isFood(isFood)
         .koreanName(koreanName)
-        .isAmbiguous(isAmbiguous)
         .build();
   }
 }
